fix 1st vs. 'other' keypresses discrimination in THE active range
- before the fix, pressing SHFT or CTRL on 'in-range' layer would also
be interpreted as in-range on layers 1..4, even though, regarding
theil key-positions they were out-of-range.

This led to the effect that the 1st actually in-range pressed key (1
out of 12) was BUBBLED only on key release instead of on key press -
if before, SHFT or CTRL was pressed.

Now with the fix:
even if SHFT or CTRL is pressed as first key on a target-layer 1..4
they are not regarded as 1st key presses (because they are not part of
the set of 12 in-range keys)