reduce required idle time for activating conditional layers
The reason being, that this did NOT allow to do fast key combination of:

- C-y
- END
(yanking to the end of line)

The reason being:
- C-y : right thumb is down (home pos)
- END : ADDITIONALLY the left thumb is down (outer pos)

The left outer thumb pressed will, as intended, first activate layer
on which RIGHT is located (even though the right thumb is still down)

The conditional layer holding END will only get activated after DELAY
time has passed.

Those 2 thumbs then activate the conditional layer containing END
If this delay is to large, i.e. if the key onto which RIGHT and END is
mapped is pressed before the DELAY has passed, then RIGHT will be sent,
else END.